turning plotfsc analysis script into a utility tool
